refactor: Move price feed of asb into swap-feed crate
Also move the tracing-subscriber deps into the workspace
