+++ monero-rpc-pool/Cargo.toml
@@ -26,7 +26,7 @@ tokio = { workspace = true, features = ["full"] }
 tower = "0.4"
 tower-http = { version = "0.5", features = ["cors"] }
 tracing = { workspace = true }
-tracing-subscriber = { version = "0.3", features = ["env-filter"] }
+tracing-subscriber = { workspace = true }
 typeshare = { workspace = true }
 url = "2.0"
 uuid = { workspace = true }
