+++ swap-feed/src/bin/kraken_ticker.rs
@@ -9,7 +9,7 @@ async fn main() -> Result<()> {
 
     let price_ticker_ws_url = Url::parse("wss://ws.kraken.com")?;
     let mut ticker =
-        swap::kraken::connect(price_ticker_ws_url).context("Failed to connect to kraken")?;
+        swap_feed::kraken::connect(price_ticker_ws_url).context("Failed to connect to kraken")?;
 
     loop {
         match ticker.wait_for_next_update().await? {
