+++ swap/src/asb.rs
@@ -4,11 +4,12 @@ mod network;
 mod rate;
 mod recovery;
 
-pub use event_loop::{EventLoop, EventLoopHandle, FixedRate, KrakenRate, LatestRate};
+pub use event_loop::{EventLoop, EventLoopHandle};
 pub use network::behaviour::{Behaviour, OutEvent};
 pub use network::rendezvous::RendezvousNode;
 pub use network::transport;
 pub use rate::Rate;
+pub use swap_feed::{FixedRate, KrakenRate, LatestRate};
 pub use recovery::cancel::cancel;
 pub use recovery::punish::punish;
 pub use recovery::redeem::{redeem, Finality};
